Migrate count array, split array and signal edge filters (#1608)(#1609)(#1610)
